// frontend/src/config/config.d.ts
import Config from "../types/config.type";
export declare enum OperationName {
    income = "income",
    expense = "expense",
    default = "default"
}
export declare enum Page {
    createCategory = "createCategory",
    editCategory = "editCategory",
    listCategories = "listCategories",
    createOperation = "createOperation",
    editOperation = "editOperation"
}
declare const config: Config;
export default config;
//# sourceMappingURL=config.d.ts.map